backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.42.0",
    "fastapi[standard]>=0.115.6",
    "openai>=1.59.8",
    "python-dotenv>=1.0.1",
]
